Refactored Component files
